combining student-matching-page into one page
